Use `gimli::write::ConvertUnit` for DWARF transform
This uses on `gimli` to provide the generic support for DWARF
transformation. Most of the remaining transform code is specific to the
needs of wasmtime.

The overall behaviour is the same as the previous implementation. We
build a graph of the DIE dependencies, and prune DIEs that don't have
valid code ranges. Then we traverse the DIE tree again and transform the
DIEs. However there are some differences.

Previously, unresolved references to DIEs were stored in `PendingUnitRefs`
and `PendingDebugInfoRefs`, then at the end of the transformation we
went back and fixed up these references. The new behaviour is to
instead reserve IDs for all of the DIEs in the dependency tree before
doing the transformation. This avoids the need to fix up references
later. It also allows us to correctly handle references in DWARF
expressions. Doing this with the previous scheme would have been
complicated. This doesn't currently matter for wasmtime because it
doesn't handle references in expressions yet, but gimli needed to handle
it in order to be correct in general.

The DIE tree pruning is slightly different. We no longer add back-edges
pointing to namespace DIEs or the root DIE. Previously, these back-edges
caused the DIEs to be reachable in cases where they should not
have been. In particular, this affects DW_TAG_variable DIEs for
global variables. We can't currently translate globals correctly,
so they were added without a valid DW_AT_location, and pulled in
references to their DW_AT_type entries. In the future when global
variables can be translated correctly, they can be included by calling
`ReserveUnitSection::require_entry`.